requirements.txt
Django==5.2.5
Pillow
django-ckeditor
